Move cmd/telliot/cli to pkg/cli
